[VCDA-1531] CLI skeleton (#606)
* CLI skeleton framework to support native and def cluster commands


* review comments addressed
